Coordinate ordering should not have an impact on the equality
The build variant model is quite important as it lay a flat layout for
the multiple mix and matching that can occurs in the native space. It
also has a lot of legacy tied into it which we are trying to remove. The
name and coordinate are tied together which this commit changes. The
name happens to follow the coordinate ordering. However the equality is
truely only tied to the coordinate. A tuple of (x, y) vs (y, x) are
equals. The coordinate always track the axis so tuple can be in any
order. The ordering is only important for reproducibility.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>