Configure JVM source set properly
We decided against bridging the Gradle SourceSet to the universal model.
Although it is technically possible, we don't believe it's the right
thing to do. If we want to allow JVM language source set, we should take
charge of the entire JVM modeling which we are not ready to do at this
point.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>